requirements.txt
pandas==2.0.3
Pillow==10.0.0
Pillow==10.2.0
scikit_learn==1.3.0
torch==2.0.1+cu118
jiwer==2.2.0
transformers==4.11.3
tqdm==4.62.3
datasets==1.12.1
```
